+++ src/OpenTelemetry.Instrumentation.EventCounters/EventCountersMetrics.cs
@@ -19,6 +19,7 @@ using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Diagnostics.Metrics;
 using System.Diagnostics.Tracing;
+using System.Globalization;
 
 namespace OpenTelemetry.Instrumentation.EventCounters;
 
@@ -29,10 +30,15 @@ internal sealed class EventCountersMetrics : EventListener
 {
     internal static readonly Meter MeterInstance = new(typeof(EventCountersMetrics).Assembly.GetName().Name, typeof(EventCountersMetrics).Assembly.GetName().Version.ToString());
 
+    private const string Prefix = "ec";
+    private const int MaxInstrumentNameLength = 63;
+
     private readonly EventCountersInstrumentationOptions options;
     private readonly List<EventSource> preInitEventSources = new();
+    private readonly List<EventSource> enabledEventSources = new();
     private readonly ConcurrentDictionary<(string, string), Instrument> instruments = new();
     private readonly ConcurrentDictionary<(string, string), double> values = new();
+    private bool isDisposed;
 
     /// <summary>
     /// Initializes a new instance of the <see cref="EventCountersMetrics"/> class.
@@ -49,6 +55,7 @@ internal sealed class EventCountersMetrics : EventListener
                 if (this.options.ShouldListenToSource(eventSource.Name))
                 {
                     this.EnableEvents(eventSource);
+                    this.enabledEventSources.Add(eventSource);
                 }
             }
 
@@ -57,10 +64,41 @@ internal sealed class EventCountersMetrics : EventListener
     }
 
     /// <inheritdoc />
+    public override void Dispose()
+    {
+        if (!this.isDisposed)
+        {
+            lock (this.preInitEventSources)
+            {
+                if (!this.isDisposed)
+                {
+                    foreach (var eventSource in this.enabledEventSources)
+                    {
+                        this.DisableEvents(eventSource);
+                    }
+
+                    this.isDisposed = true;
+                }
+            }
+
+            // DO NOT clear the ConcurrentDictionary instances as some other thread executing the OnEventWritten callback might be using them
+            this.enabledEventSources.Clear();
+            this.options.EventSourceNames.Clear();
+        }
+
+        base.Dispose();
+    }
+
+    /// <inheritdoc />
     protected override void OnEventSourceCreated(EventSource eventSource)
     {
         lock (this.preInitEventSources)
         {
+            if (this.isDisposed)
+            {
+                return;
+            }
+
             if (this.options == null)
             {
                 this.preInitEventSources.Add(eventSource);
@@ -68,6 +106,7 @@ internal sealed class EventCountersMetrics : EventListener
             else if (this.options.ShouldListenToSource(eventSource.Name))
             {
                 this.EnableEvents(eventSource);
+                this.enabledEventSources.Add(eventSource);
             }
         }
     }
@@ -113,12 +152,45 @@ internal sealed class EventCountersMetrics : EventListener
             return;
         }
 
-        var value = Convert.ToDouble(hasMean ? mean : increment);
+        var value = Convert.ToDouble(hasMean ? mean : increment, CultureInfo.InvariantCulture);
         this.UpdateInstrumentWithEvent(hasMean, eventSourceName, name, value);
     }
 
     private static Dictionary<string, string> GetEnableEventsArguments(EventCountersInstrumentationOptions options) =>
-        new() { { "EventCounterIntervalSec", options.RefreshIntervalSecs.ToString() } };
+        new() { { "EventCounterIntervalSec", options.RefreshIntervalSecs.ToString(CultureInfo.InvariantCulture) } };
+
+    /// <summary>
+    /// If the resulting instrument name is too long, it trims the event source name
+    /// to fit in as many characters as possible keeping the event name intact.
+    /// E.g. instrument for `Microsoft-AspNetCore-Server-Kestrel`, `tls-handshakes-per-second`
+    /// would be too long (64 chars), so it's shortened to `ec.Microsoft-AspNetCore-Server-Kestre.tls-handshakes-per-second`.
+    ///
+    /// If there is no room for event source name, returns `ec.{event name}` and
+    /// if it's still too long, it will be validated and ignored later in the pipeline.
+    /// </summary>
+    private static string GetInstrumentName(string sourceName, string eventName)
+    {
+        int totalLength = Prefix.Length + 1 + sourceName.Length + 1 + eventName.Length;
+        if (totalLength <= MaxInstrumentNameLength)
+        {
+            return string.Concat(Prefix, ".", sourceName, ".", eventName);
+        }
+
+        var maxEventSourceLength = MaxInstrumentNameLength - Prefix.Length - 2 - eventName.Length;
+        if (maxEventSourceLength < 1)
+        {
+            // event name is too long, there is not enough space for sourceName.
+            // let ec.<eventName> flow to metrics SDK and it will suppress it if needed.
+            return string.Concat(Prefix, ".", eventName);
+        }
+
+        while (maxEventSourceLength > 0 && (sourceName[maxEventSourceLength - 1] == '.' || sourceName[maxEventSourceLength - 1] == '-'))
+        {
+            maxEventSourceLength--;
+        }
+
+        return string.Concat(Prefix, ".", sourceName.Substring(0, maxEventSourceLength), ".", eventName);
+    }
 
     private void EnableEvents(EventSource eventSource)
     {
@@ -132,10 +204,9 @@ internal sealed class EventCountersMetrics : EventListener
             ValueTuple<string, string> metricKey = new(eventSourceName, name);
             _ = this.values.AddOrUpdate(metricKey, value, isGauge ? (_, _) => value : (_, existing) => existing + value);
 
-            var instrumentName = $"EventCounters.{eventSourceName}.{name}";
-
             if (!this.instruments.ContainsKey(metricKey))
             {
+                var instrumentName = GetInstrumentName(eventSourceName, name);
                 Instrument instrument = isGauge
                     ? MeterInstance.CreateObservableGauge(instrumentName, () => this.values[metricKey])
                     : MeterInstance.CreateObservableCounter(instrumentName, () => this.values[metricKey]);