+++ test/OpenTelemetry.Exporter.Geneva.Benchmark/Exporter/TraceExporterBenchmarks.cs
@@ -14,24 +14,24 @@
 // limitations under the License.
 // </copyright>
 
-using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using BenchmarkDotNet.Attributes;
 using OpenTelemetry.Trace;
 
 /*
-BenchmarkDotNet=v0.13.1, OS=Windows 10.0.22621
+BenchmarkDotNet=v0.13.2, OS=Windows 11 (10.0.22621.963)
 Intel Core i7-9700 CPU 3.00GHz, 1 CPU, 8 logical and 8 physical cores
-.NET SDK=7.0.100-preview.6.22352.1
-  [Host]     : .NET 6.0.9 (6.0.922.41905), X64 RyuJIT
-  DefaultJob : .NET 6.0.9 (6.0.922.41905), X64 RyuJIT
+.NET SDK=7.0.101
+  [Host]     : .NET 7.0.1 (7.0.122.56804), X64 RyuJIT AVX2
+  DefaultJob : .NET 7.0.1 (7.0.122.56804), X64 RyuJIT AVX2
 
 
-|            Method |     Mean |   Error |  StdDev |  Gen 0 | Allocated |
-|------------------ |---------:|--------:|--------:|-------:|----------:|
-|    ExportActivity | 719.9 ns | 5.83 ns | 5.45 ns | 0.0057 |      40 B |
-| SerializeActivity | 361.7 ns | 1.09 ns | 0.97 ns | 0.0062 |      40 B |
+|                           Method |     Mean |    Error |   StdDev |   Median |   Gen0 | Allocated |
+|--------------------------------- |---------:|---------:|---------:|---------:|-------:|----------:|
+|                   ExportActivity | 566.3 ns |  3.13 ns |  2.44 ns | 565.9 ns |      - |         - |
+|                SerializeActivity | 313.3 ns |  1.71 ns |  1.60 ns | 313.0 ns |      - |         - |
+| CreateActivityWithGenevaExporter | 940.5 ns | 18.77 ns | 54.14 ns | 911.4 ns | 0.0648 |     416 B |
 */
 
 namespace OpenTelemetry.Exporter.Geneva.Benchmark;
@@ -42,6 +42,7 @@ public class TraceExporterBenchmarks
     private readonly Activity activity;
     private readonly Batch<Activity> batch;
     private readonly MsgPackTraceExporter exporter;
+    private readonly TracerProvider tracerProvider;
     private readonly ActivitySource activitySource = new ActivitySource("OpenTelemetry.Exporter.Geneva.Benchmark");
 
     public TraceExporterBenchmarks()
@@ -50,7 +51,8 @@ public class TraceExporterBenchmarks
 
         this.batch = this.CreateBatch();
 
-        using var activityListener = new ActivityListener
+        #region Create activity to be used for Serialize and Export benchmark methods
+        var activityListener = new ActivityListener
         {
             ActivityStarted = null,
             ActivityStopped = null,
@@ -63,10 +65,13 @@ public class TraceExporterBenchmarks
         using (var testActivity = this.activitySource.StartActivity("Benchmark"))
         {
             this.activity = testActivity;
-            this.activity?.SetTag("tagString", "value");
-            this.activity?.SetTag("tagInt", 100);
-            this.activity?.SetStatus(Status.Error);
+            this.activity.SetTag("tagString", "value");
+            this.activity.SetTag("tagInt", 100);
+            this.activity.SetStatus(Status.Error);
         }
+
+        activityListener.Dispose();
+        #endregion
 
         this.exporter = new MsgPackTraceExporter(new GenevaExporterOptions
         {
@@ -78,6 +83,21 @@ public class TraceExporterBenchmarks
                 ["cloud.roleVer"] = "9.0.15289.2",
             },
         });
+
+        this.tracerProvider = Sdk.CreateTracerProviderBuilder()
+            .SetSampler(new AlwaysOnSampler())
+            .AddSource(this.activitySource.Name)
+            .AddGenevaTraceExporter(options =>
+            {
+                options.ConnectionString = "EtwSession=OpenTelemetry";
+                options.PrepopulatedFields = new Dictionary<string, object>
+                {
+                    ["cloud.role"] = "BusyWorker",
+                    ["cloud.roleInstance"] = "CY1SCH030021417",
+                    ["cloud.roleVer"] = "9.0.15289.2",
+                };
+            })
+            .Build();
     }
 
     [Benchmark]
@@ -92,6 +112,13 @@ public class TraceExporterBenchmarks
         this.exporter.SerializeActivity(this.activity);
     }
 
+    [Benchmark]
+    public void CreateActivityWithGenevaExporter()
+    {
+        // this activity will be created and sent to Geneva exporter
+        using var activity = this.activitySource.StartActivity("Benchmark");
+    }
+
     [GlobalCleanup]
     public void Cleanup()
     {
@@ -99,6 +126,7 @@ public class TraceExporterBenchmarks
         this.activitySource.Dispose();
         this.batch.Dispose();
         this.exporter.Dispose();
+        this.tracerProvider.Dispose();
     }
 
     private Batch<Activity> CreateBatch()