+++ test/OpenTelemetry.Instrumentation.EventCounters.Tests/EventCountersMetricsTests.cs
@@ -28,33 +28,33 @@ public class EventCountersMetricsTests
     private const int Delay = 1200;
 
     [Fact(Skip = "Other tests metrics are being exported here")]
-    public async Task NoMetricsByDefault()
+    public void NoMetricsByDefault()
     {
         // Arrange
         List<Metric> metricItems = new();
 
-        var meterProvider = Sdk.CreateMeterProviderBuilder()
-            .AddEventCountersInstrumentation()
+        using var meterProvider = Sdk.CreateMeterProviderBuilder()
+            .AddEventCountersInstrumentation(null)
             .AddInMemoryExporter(metricItems)
             .Build();
 
         // Act
-        await Task.Delay(Delay);
+        Task.Delay(Delay).Wait();
         meterProvider.ForceFlush();
 
         // Assert
         Assert.Empty(metricItems);
     }
 
-    [Fact(Skip = "Unstable")]
-    public async Task EventCounter()
+    [Fact]
+    public void EventCounter()
     {
         // Arrange
         List<Metric> metricItems = new();
-        EventSource source = new("a");
-        EventCounter counter = new("c", source);
+        using EventSource source = new("a");
+        using EventCounter counter = new("c", source);
 
-        var meterProvider = Sdk.CreateMeterProviderBuilder()
+        using var meterProvider = Sdk.CreateMeterProviderBuilder()
             .AddEventCountersInstrumentation(options =>
             {
                 options.AddEventSources(source.Name);
@@ -64,25 +64,25 @@ public class EventCountersMetricsTests
 
         // Act
         counter.WriteMetric(1997.0202);
-        await Task.Delay(Delay);
+        Task.Delay(Delay).Wait();
         meterProvider.ForceFlush();
 
         // Assert
-        var metric = metricItems.Find(x => x.Name == "EventCounters.a.c");
+        var metric = metricItems.Find(x => x.Name == "ec.a.c");
         Assert.NotNull(metric);
         Assert.Equal(MetricType.DoubleGauge, metric.MetricType);
         Assert.Equal(1997.0202, GetActualValue(metric));
     }
 
     [Fact]
-    public async Task IncrementingEventCounter()
+    public void IncrementingEventCounter()
     {
         // Arrange
         List<Metric> metricItems = new();
-        EventSource source = new("b");
-        IncrementingEventCounter incCounter = new("inc-c", source);
+        using EventSource source = new("b");
+        using IncrementingEventCounter incCounter = new("inc-c", source);
 
-        var meterProvider = Sdk.CreateMeterProviderBuilder()
+        using var meterProvider = Sdk.CreateMeterProviderBuilder()
             .AddEventCountersInstrumentation(options =>
             {
                 options.AddEventSources(source.Name);
@@ -94,26 +94,26 @@ public class EventCountersMetricsTests
         incCounter.Increment(1);
         incCounter.Increment(1);
         incCounter.Increment(1);
-        await Task.Delay(Delay);
+        Task.Delay(Delay).Wait();
         meterProvider.ForceFlush();
 
         // Assert
-        var metric = metricItems.Find(x => x.Name == "EventCounters.b.inc-c");
+        var metric = metricItems.Find(x => x.Name == "ec.b.inc-c");
         Assert.NotNull(metric);
         Assert.Equal(MetricType.DoubleSum, metric.MetricType);
         Assert.Equal(3, GetActualValue(metric));
     }
 
     [Fact(Skip = "Unstable")]
-    public async Task PollingCounter()
+    public void PollingCounter()
     {
         // Arrange
         int i = 0;
         List<Metric> metricItems = new();
-        EventSource source = new("c");
-        PollingCounter pollCounter = new("poll-c", source, () => ++i * 10);
+        using EventSource source = new("c");
+        using PollingCounter pollCounter = new("poll-c", source, () => ++i * 10);
 
-        var meterProvider = Sdk.CreateMeterProviderBuilder()
+        using var meterProvider = Sdk.CreateMeterProviderBuilder()
             .AddEventCountersInstrumentation(options =>
             {
                 options.AddEventSources(source.Name);
@@ -122,26 +122,26 @@ public class EventCountersMetricsTests
             .Build();
 
         // Act
-        await Task.Delay(Delay * 2);
+        Task.Delay(Delay * 2).Wait();
         meterProvider.ForceFlush();
 
         // Assert
-        var metric = metricItems.Find(x => x.Name == "EventCounters.c.poll-c");
+        var metric = metricItems.Find(x => x.Name == "ec.c.poll-c");
         Assert.NotNull(metric);
         Assert.Equal(MetricType.DoubleGauge, metric.MetricType);
         Assert.Equal(20, GetActualValue(metric));
     }
 
     [Fact(Skip = "Unstable")]
-    public async Task IncrementingPollingCounter()
+    public void IncrementingPollingCounter()
     {
         // Arrange
         int i = 1;
         List<Metric> metricItems = new();
-        EventSource source = new("d");
-        IncrementingPollingCounter incPollCounter = new("inc-poll-c", source, () => i++);
+        using EventSource source = new("d");
+        using IncrementingPollingCounter incPollCounter = new("inc-poll-c", source, () => i++);
 
-        var meterProvider = Sdk.CreateMeterProviderBuilder()
+        using var meterProvider = Sdk.CreateMeterProviderBuilder()
             .AddEventCountersInstrumentation(options =>
             {
                 options.AddEventSources(source.Name);
@@ -150,26 +150,26 @@ public class EventCountersMetricsTests
             .Build();
 
         // Act
-        await Task.Delay(Delay * 2);
+        Task.Delay(Delay * 2).Wait();
         meterProvider.ForceFlush();
 
         // Assert
-        var metric = metricItems.Find(x => x.Name == "EventCounters.d.inc-poll-c");
+        var metric = metricItems.Find(x => x.Name == "ec.d.inc-poll-c");
         Assert.NotNull(metric);
         Assert.Equal(MetricType.DoubleSum, metric.MetricType);
         Assert.Equal(2, GetActualValue(metric));
     }
 
-    [Fact(Skip = "Unstable")]
-    public async Task EventCounterSameNameUsesNewestCreated()
+    [Fact]
+    public void EventCounterSameNameUsesNewestCreated()
     {
         // Arrange
         List<Metric> metricItems = new();
-        EventSource source = new("a");
-        EventCounter counter = new("c", source);
-        EventCounter counter2 = new("c", source);
+        using EventSource source = new("a");
+        using EventCounter counter = new("c", source);
+        using EventCounter counter2 = new("c", source);
 
-        var meterProvider = Sdk.CreateMeterProviderBuilder()
+        using var meterProvider = Sdk.CreateMeterProviderBuilder()
             .AddEventCountersInstrumentation(options =>
             {
                 options.AddEventSources(source.Name);
@@ -180,11 +180,11 @@ public class EventCountersMetricsTests
         // Act
         counter2.WriteMetric(1980.1208);
         counter.WriteMetric(1997.0202);
-        await Task.Delay(Delay);
+        Task.Delay(Delay).Wait();
         meterProvider.ForceFlush();
 
         // Assert
-        var metric = metricItems.Find(x => x.Name == "EventCounters.a.c");
+        var metric = metricItems.Find(x => x.Name == "ec.a.c");
         Assert.NotNull(metric);
         Assert.Equal(MetricType.DoubleGauge, metric.MetricType);
 
@@ -212,6 +212,73 @@ public class EventCountersMetricsTests
         });
 
         Assert.Equal("Use the `OpenTelemetry.Instrumentation.Runtime` or `OpenTelemetry.Instrumentation.Process` instrumentations.", ex.Message);
+    }
+
+    [Theory]
+    [InlineData("Microsoft-AspNetCore-Server-Kestrel-1", "tls-handshakes-per-second", "ec.Microsoft-AspNetCore-Server-Kestre.tls-handshakes-per-second")]
+    [InlineData("Microsoft-AspNetCore-Server-Kestrel-1", "tls-handshakes-per-sec", "ec.Microsoft-AspNetCore-Server-Kestrel-1.tls-handshakes-per-sec")]
+    [InlineData("Microsoft.AspNetCore.Http.Connections-1", "connections-stopped", "ec.Microsoft.AspNetCore.Http.Connections-1.connections-stopped")]
+    [InlineData("Microsoft.AspNetCore.Http.Connections-1", "connections-timed-out-longer", "ec.Microsoft.AspNetCore.Http.Conne.connections-timed-out-longer")]
+    [InlineData("Microsoft.AspNetCore.Http.Conn.Something", "connections-timed-out-longer", "ec.Microsoft.AspNetCore.Http.Conn.connections-timed-out-longer")]
+    [InlineData("Microsoft.AspNetCore.One.Two", "very-very-very-very-very-very-very-very-very-long-event-name", "ec.very-very-very-very-very-very-very-very-very-long-event-name")]
+    [InlineData("Microsoft.AspNetCore.One.Two", "very-very-very-very-very-very-very-very-long-event-name", "ec.Micr.very-very-very-very-very-very-very-very-long-event-name")]
+    [InlineData("Microsoft.AspNetCore.One.Two", "very-very-very-very-very-very-very-long-event-name", "ec.Microsoft.very-very-very-very-very-very-very-long-event-name")]
+    public void EventSourceNameShortening(string sourceName, string eventName, string expectedInstrumentName)
+    {
+        // Arrange
+        List<Metric> metricItems = new();
+        using EventSource source = new(sourceName);
+        using IncrementingEventCounter connections = new(eventName, source);
+
+        using var meterProvider = Sdk.CreateMeterProviderBuilder()
+            .AddEventCountersInstrumentation(options =>
+            {
+                options.AddEventSources(source.Name);
+            })
+            .AddInMemoryExporter(metricItems)
+            .Build();
+
+        // Act
+        connections.Increment(1);
+        Task.Delay(Delay).Wait();
+        meterProvider.ForceFlush();
+
+        // Assert
+        Metric metric = metricItems.Find(m => m.Name == expectedInstrumentName);
+        Assert.NotNull(metric);
+        Assert.Equal(1, GetActualValue(metric));
+    }
+
+    [Fact]
+    public void InstrumentNameTooLong()
+    {
+        // Arrange
+        List<Metric> metricItems = new();
+        using EventSource source = new("source");
+
+        // ec.s. + event name is 63;
+        string veryLongEventName = new string('e', 100);
+        using IncrementingEventCounter connections = new(veryLongEventName, source);
+
+        using var meterProvider = Sdk.CreateMeterProviderBuilder()
+            .AddEventCountersInstrumentation(options =>
+            {
+                options.AddEventSources(source.Name);
+            })
+            .AddInMemoryExporter(metricItems)
+            .Build();
+
+        // Act
+        connections.Increment(1);
+        Task.Delay(Delay).Wait();
+        meterProvider.ForceFlush();
+
+        // Assert
+        foreach (var item in metricItems)
+        {
+            Assert.False(item.Name.StartsWith("ec.source.ee"));
+            Assert.False(item.Name.StartsWith("ec.s.ee"));
+        }
     }
 
     // polling and eventcounter with same instrument name?