+++ test/OpenTelemetry.Instrumentation.Runtime.Tests/RuntimeMetricsTests.cs
@@ -75,6 +75,9 @@ public class RuntimeMetricsTests
         var gcCountMetric = exportedItems.FirstOrDefault(i => i.Name == "process.runtime.dotnet.gc.collections.count");
         Assert.NotNull(gcCountMetric);
 
+        var totalObjectsSize = exportedItems.FirstOrDefault(i => i.Name == "process.runtime.dotnet.gc.objects.size");
+        Assert.NotNull(totalObjectsSize);
+
 #if NET6_0_OR_GREATER
 
         var gcAllocationSizeMetric = exportedItems.FirstOrDefault(i => i.Name == "process.runtime.dotnet.gc.allocations.size");
@@ -130,7 +133,7 @@ public class RuntimeMetricsTests
         List<Task> tasks = new List<Task>();
         for (int i = 0; i < taskCount; i++)
         {
-            tasks.Add(Task.Run(() => { Console.Write("Hi"); }));
+            tasks.Add(Task.Run(() => { }));
         }
 
         Task.WaitAll(tasks.ToArray());