+++ src/OpenTelemetry.Exporter.Geneva/CHANGELOG.md
@@ -2,17 +2,52 @@
 
 ## Unreleased
 
+## 1.4.0-rc.1
+
+Released 2022-Dec-19
+
+* Update OpenTelemetry to 1.4.0-rc.1
+  ([#820](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/820))
+* Add support in logs for prefix-based table name mapping configuration.
+  [#796](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/796)
+* Updated the trace exporter to use the new performance APIs introduced in
+  `System.Diagnostics.DiagnosticSource` v7.0.
+  [#838](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/838)
+* Avoid allocation when serializing scopes.
+  ([#818](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/818))
+
+## 1.4.0-beta.6
+
+Released 2022-Dec-09
+
+* Added support for
+  [DateTimeOffset](https://learn.microsoft.com/dotnet/api/system.datetimeoffset).
+  ([#797](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/797))
+* Fix the overflow bucket value serialization for Histogram.
+  ([#805](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/805))
+* Fix EventSource logging.
+  ([#813](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/813))
+
+## 1.4.0-beta.5
+
+Released 2022-Nov-21
+
+* Update OpenTelemetry to 1.4.0-beta.3
+  ([#774](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/774))
+
 ## 1.4.0-beta.4
 
 Released 2022-Oct-28
 
-* Breaking change: Updated export logic for scopes
+* Updated export logic for scopes
+  * Users upgrading from `1.4.0-beta.1`, `1.4.0-beta.2` or `1.4.0-beta.3` to
+    this version will see a **breaking change**
   * Export scopes which have a non-null key as individual columns (each
     key-value pair from the scopes is exported as its own column; these columns
     would also be taken into consideration when the CustomFields option is
     applied).
-  * When using formatted strings for scopes, the templated string (`"{OriginalFormat"}`)
-    will not be exported.
+  * When using formatted strings for scopes, the templated string
+    (`"{OriginalFormat"}`) will not be exported.
   [#736](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/736)
 
 ## 1.4.0-beta.3