+++ src/OpenTelemetry.Instrumentation.Wcf/CHANGELOG.md
@@ -2,8 +2,14 @@
 
 ## Unreleased
 
+## 1.0.0-rc.8
+
+Released 2022-Dec-28
+
 * Update OTel SDK version to `1.3.1`.
   ([#631](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/631))
+* Change value `rpc.system` from `wcf` to `dotnet_wcf`.
+  ([#837](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/837))
 
 ## 1.0.0-rc.7